requirements.txt
# ────────────────────────────────────────────────
# Agentic AI Travel Planner  ·  Requirements
# ────────────────────────────────────────────────

# ── Core Framework ─────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
flask>=3.0.0
pydantic>=2.6.0

# ── LangChain / LangGraph ─────────────────────
langchain>=0.3.0
langchain-core>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# ── Database ───────────────────────────────────
SQLAlchemy>=2.0.0

# ── PDF Generation ─────────────────────────────
xhtml2pdf>=0.2.0
markdown>=3.5.0

# ── API & Networking ──────────────────────────
requests>=2.31.0

# ── Configuration & Environment ────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0
